LUT ranges

+++ src/dsp_utils/lut.h
@@ -8,8 +8,8 @@ extern "C"
 #endif
 
     extern const float cos_lut[];
-    extern const int8_t cos_lut_fixed8[];
-    extern const int16_t cos_lut_fixed16[];
+    extern const int16_t cos_lut_fixed8[];
+    extern const int32_t cos_lut_fixed16[];
 
 #ifdef __cplusplus
 }
